Render map centered on the user's current location
==============================
  What?
==============================

A geolocation app needs a map for a user to see and, perhaps,
to interact with. This commit will create and render a
fullscreen map. The map will show the user's position as a
blue dot and will follow the user.

==============================
  How?
==============================

I imported a MapView component from react-native-maps to render
a map.

Installation is pretty simple:

  $ npm install react-native-maps --save
  $ react-native link

I also added styles for the map because by default, it will
render a map with dimensions of 0x0. Now, it currently fills
the entire screen.

Some attributes were added for further customization:
  - showsUserLocation={ true }
    > This will show the user's location on the map

  - followsUserLocation={ true }
    > The map will try to follow and center the map
      on the user's location

  - loadingEnabled={ true }
    > Shows a loading indicator while the map is loading

  - showsPointsOfInterest={ false }
    > Hides all points of interests like buildings,
      restaurants, entertainment areas, etc...