Remove Riddles Schema
Original purpose of riddles schema was to create unique subdoc keys for riddles inside the user schema. Unfortunately, after researching this feature it appears that while mongoosejs does allow subdocs to create 'unique' keys, when processing a subdoc as part of a larger schema those keys are ignored and will not enforce uniqueness. As such, the schema has been consolidated into the User schema and duplicate checking was added (in a previous commit) to the controller function for adding solved riddles.